requirements.txt
strands-agents
python_dotenv
litellm